% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerProcessNetCdf-package.R
\docType{package}
\name{cerProcessNetCdf-package}
\alias{cerProcessNetCdf}
\alias{cerProcessNetCdf-package}
\title{cerProcessNetCdf: Deal with NetCDF files}
\description{
Deal with NetCDF files.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
