% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridFormat.R
\name{createGridFormat}
\alias{createGridFormat}
\title{GridFormat class}
\usage{
createGridFormat(nLon, nLat, lonIncreasing, latIncreasing, lonFirst)
}
\arguments{
\item{nLon, nLat}{Number of longitude and latitude values.}

\item{lonIncreasing, latIncreasing}{Whether the longitude/latitude values are in increasing order.}

\item{lonFirst}{Whether the longitude is the first dimension.}
}
\description{
A class to store the format of a longitude - latitude grid of the globe.
}
\examples{
createGridFormat(20, 10, TRUE, TRUE, TRUE)
}
