% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessNetCdf-package.R
\docType{package}
\name{ProcessNetCdf-package}
\alias{ProcessNetCdf}
\alias{ProcessNetCdf-package}
\title{ProcessNetCdf: Deal with NetCDF files (formerly EbmNetCdf)}
\description{
Deal with NetCDF files.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
