% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popWeightAggregation.R
\name{setupPopWeightAggregation}
\alias{setupPopWeightAggregation}
\title{Setup for runPopWeightAggregation().}
\usage{
setupPopWeightAggregation(
  targetFormat,
  maskFilePath,
  maskSumFilePath,
  boundingBoxFilePath,
  popDataDescriptor = NULL,
  invarDir,
  invarFileNamePattern,
  invarDimensionName,
  invarValueVariableName,
  lonLatVarToDimOrder = 1:3,
  outDir,
  outNcFilePattern,
  batchSize
)
}
\arguments{
\item{targetFormat}{An object of class GridFormat as created by createGridFormat().}
}
\description{
Setup for runPopWeightAggregation().
}
